+++ code/test/sixsq/nuvla/server/resources/nuvlabox/utils_test.clj
@@ -7,3 +7,10 @@
 (deftest check-short-nb-id
   (is (nil? (t/short-nb-id nil)))
   (is (= "abc" (t/short-nb-id "nuvlabox/abc-def-ghi-jkl"))))
+
+
+(deftest check-has-pull-support?
+  (is (false? (t/has-pull-support? {:capabilities []})))
+  (is (false? (t/has-pull-support? {})))
+  (is (false? (t/has-pull-support? {:capabilities ["ANYTHING"]})))
+  (is (t/has-pull-support? {:capabilities ["ANYTHING" "NUVLA_JOB_PULL"]})))