+++ code/test/sixsq/nuvla/server/resources/nuvlabox_2_lifecycle_test.clj
@@ -62,7 +62,7 @@
                      :form-factor      "Nuvlabox"
                      :lan-cidr         "10.0.1.0/24"
                      :ssh-keys         ["credential/aaa-bbb-ccc"]
-                     :capabilities     ["RANDOM"]})
+                     :capabilities     ["RANDOM" "NUVLA_JOB_PULL"]})
 
 
 (deftest check-metadata
@@ -1156,7 +1156,7 @@
           (ltu/body->edn)
           (ltu/is-status 200))
 
-      ;; push
+      ;; pull
       (let [reboot-url (-> session-owner
                            (request nuvlabox-url)
                            (ltu/body->edn)
@@ -1169,14 +1169,6 @@
                            (ltu/location-url))]
         (-> session-admin
             (request job-url)
-            (ltu/body->edn)
-            (ltu/is-status 200)
-            (ltu/is-key-value :execution-mode "push"))
-
-        (-> session-admin
-            (request job-url
-                     :request-method :put
-                     :body (json/write-str {:execution-mode "pull"}))
             (ltu/body->edn)
             (ltu/is-status 200)
             (ltu/is-key-value :execution-mode "pull")